chore(payment): PAYMENTS-7402 rename ppsdk payment-processor files to sub-strategy
