+++ src/payment/strategies/ppsdk/index.ts
@@ -1,4 +1,4 @@
 export { PPSDKStrategy } from './ppsdk-strategy';
-export { createPaymentProcessorRegistry } from './create-ppsdk-payment-processor-registry';
+export { createPaymentProcessorRegistry } from './create-ppsdk-sub-strategy-registry';
 export { createStepHandler } from './step-handler';
 export { PaymentResumer } from './ppsdk-payment-resumer';
