--- src/payment/strategies/ppsdk/payment-processors/index.ts
@@ -1,1 +1,0 @@
-export { NonePaymentProcessor } from './none-payment-processor';
