+++ src/payment/strategies/ppsdk/sub-strategies/index.ts
@@ -1,0 +1,1 @@
+export { NonePaymentProcessor } from './none-sub-strategy';
